+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+2010-03-08
+==========
+* Fixed the bugs
+* Made the extra columns read-only.
+
 2010-03-07
 ==========
 * First release